Enable rectangular partial screen updates
Add the ability to group scanline updates into rectangular updates and use
write-continue. This avoids a lot of overhead with setting page/column
addresses and allows us to go back to sending multiple lines at a time.

This is quite a big boost to performance and has some headroom to tweak
and improve.