Improve frame-skip, fix SMS/GG buffer diff and tweak display defaults
Make the frame-skip measure time taken before the audio buffer so it
doesn't unnecessarily skip frames when we're keeping up. This allows us to
hit 60Hz on simple or static-screen games on the NES quite easily, but
really makes 30Hz drops stand out. Probably need to revisit interlacing.

Fixed buffer diff code to use palettes - this fixes SMS/GG buffer diffing
as the palette can change between frames (and though the SMS only has a
6-bit possible palette, the GG has 12-bit, so we can't just generate a
single palette and have done with it).

Tweaked partial update defaults so that the minimum interrupt-based update
length is now the size of a line buffer. Also tweaked the write_rect
function to fill the buffer as much as possible, rather than taking the
line-count and ignoring line-width. There are still cases, mainly on
scaled SMS for some reason, where the partial-update path is slower than
just updating everything, so more debugging is necessary.