Put more screen code in IRAM, improve debugging output on NES
Currently, a busy update (e.g. Contra) causes us to output at ~37-43fps.
Forcing a full interlaced update puts us at about 47fps, which is an
improvement, but given the visual degradation isn't really worth it imho.
All these figures are for an unscaled screen.

I'm still looking at ways to attain 60Hz interlaced updates. If we allow
the screen update code to run in parallel to the emulator processing the
next frame, we get an improved frame-rate, but it seems to conflict somehow
with the screen update and looks terrible, at least with polled updates.
Getting that to work may be the key to 60Hz interlaced updates, given the
tight timings.

Note that later esp-idf checkouts allow explicitly marking all SPI code as
being in IRAM and stops FreeRTOS from disabling flash, which may improve
code throughput. However, using a later esp-idf does something weird where
when we aren't updating the screen, the sound crackles. Almost like
the SPI traffic being some kind of keep-alive that's affecting the audio
(which I assume is also on the SPI bus?)