Integrate interlaced updates as a fall-back when we skip too many frames
I'm still not pleased with the bottom level of performance, which seems in
the worst case to be about 44/45fps, but this is certainly better than it
is without any of these patches.

In theory, we're uploading to the screen in parallel now, but I'm not sure
I believe it's happening, I think it might be getting blocked by audio
updates. The emulation takes about 4ms, which in a perfect situation, ought
to give us a 4ms head-start in rendering, which really ought to boost our
bottom level of performance from where it is to around 55fps... Further
investigation is necessary. Perhaps we could off-load input and audio
updates to the IO task and rule out the possibility of there being a
conflict.