Simplify send_reset_drawing / send_continue_line some
It doesn't mention it explicitly in the documentation, but setting the page
and column addresses accepts fewer than 4 parameters, so we can save on
some SPI bandwidth by not setting the end addresses when it's unnecessary.

I also noticed that send_continue_line was sending an unnecessary
write-continue for each line buffer, so that cuts down there also. Turns
out there's no real need for a separate send_write_line function as
whether or not you're going to continue can be determined from the height
given to send_reset_drawing, so that function has been removed.