Tweak interlace thresholds for better frame-rates
So the last SPI traffic optimisation actually had a dramatic effect, but
it was hidden by the chosen interlace thresholds. Now frame-rates of
unscaled busy Contra on the first level vary from about 49-60, with the
majority of busy frames over 50fps. Interlacing forced on is 56-60, forced
off is 40-60. Things are looking much better now :)