nes: Restore old vidQueue and make states use a union
Restore the old queue method for handling video tasks. Notification is
quicker and uses substantially less memory, but also limits what can be
done in the task if you want to use queues/semaphores. I don't think
there's any really substantial performance benefit.

That unfortunately pushed memory use up a little bit to where saving
states was no longer reliable (and reports of some games had it already
being unreliable), but there was an easy win there - the save state
structure houses copies of all the data needed to save a NES state, but
it only saves it part-by-part. There was no need to have this all
allocated at once, so making each block live in a union saves a lot of
memory.

An easy win if they're reported to still be unreliable would be to just
not copy that memory into a separate structure and save it directly from
the machine structures.